requirements.txt
appdirs==1.4.3
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.3
bleach==2.0.0
Bottleneck==1.2.0
certifi==2017.1.23
coverage==4.0.1
cycler==0.10.0
Cython==0.25.2
decorator==4.0.11
docopt==0.6.2
docutils==0.13.1
entrypoints==0.2.2
et-xmlfile==1.0.1
funcsigs==1.0.2
google-apputils==0.4.1
h5py==2.6.0
html5lib==0.999999999
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jdcal==1.0
jedi==0.10.2
Jinja2==2.9.5
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.0.0
jupyter-core==4.3.0
jupyterlab==0.4.1
Keras==2.0.3
lxml==3.7.2
MarkupSafe==1.0
matplotlib==2.0.0
mistune==0.7.1
mock==2.0.0
more-itertools==3.0.0
mox==0.5.3
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nose==1.3.7
notebook==5.7.2
numexpr==2.6.2
numpy==1.11.3
openpyxl==2.3.5
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
path.py==10.1
pbr==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
pkgconfig==1.1.0
prompt-toolkit==1.0.14
protobuf==3.2.0
ptyprocess==0.5
py==1.4.31
pydot-ng==1.0.0
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-gflags==2.0
pytz==2016.6.1
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.20.0
scikit-learn==0.18.1
scipy==0.18.1
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.0.15
tables==3.3.0
tensorflow==1.0.0
terminado==0.6
testpath==0.3
Theano==0.8.2
thespian==3.7.2
tornado==4.4.2
traitlets==4.3.2
virtualenv==15.0.3
wcwidth==0.1.6
webencodings==0.5
widgetsnbextension==2.0.0
xlrd==0.9.4
xlwt==1.1.2
